add full dtos to dto lawyer
